fix(preview): trim trailing terminal url punctuation
